Extracted setting registering overloads out of AbstractConfig
