+++ src/main/java/me/zeroeightsix/kami/setting/GenericConfig.kt
@@ -1,7 +1,7 @@
 package me.zeroeightsix.kami.setting
 
 import me.zeroeightsix.kami.KamiMod
-import me.zeroeightsix.kami.setting.config.AbstractConfig
+import me.zeroeightsix.kami.setting.configs.AbstractConfig
 import me.zeroeightsix.kami.setting.settings.AbstractSetting
 
 internal object GenericConfig : AbstractConfig<Any>(
