+++ src/main/java/me/zeroeightsix/kami/setting/GuiConfig.kt
@@ -2,7 +2,7 @@ package me.zeroeightsix.kami.setting
 
 import me.zeroeightsix.kami.KamiMod
 import me.zeroeightsix.kami.gui.rgui.Component
-import me.zeroeightsix.kami.setting.config.AbstractMultiConfig
+import me.zeroeightsix.kami.setting.configs.AbstractMultiConfig
 import me.zeroeightsix.kami.setting.settings.AbstractSetting
 import java.io.File
