+++ src/main/java/me/zeroeightsix/kami/setting/ModuleConfig.kt
@@ -3,7 +3,7 @@ package me.zeroeightsix.kami.setting
 import me.zeroeightsix.kami.KamiMod
 import me.zeroeightsix.kami.module.Module
 import me.zeroeightsix.kami.setting.GenericConfig.setting
-import me.zeroeightsix.kami.setting.config.AbstractConfig
+import me.zeroeightsix.kami.setting.configs.AbstractConfig
 import me.zeroeightsix.kami.setting.settings.AbstractSetting
 import java.io.File
