+++ src/main/java/me/zeroeightsix/kami/setting/configs/AbstractConfig.kt
@@ -1,0 +1,77 @@
+package me.zeroeightsix.kami.setting.configs
+
+import com.google.gson.Gson
+import com.google.gson.GsonBuilder
+import com.google.gson.JsonParser
+import me.zeroeightsix.kami.KamiMod
+import me.zeroeightsix.kami.setting.groups.SettingGroup
+import me.zeroeightsix.kami.setting.groups.SettingMultiGroup
+import me.zeroeightsix.kami.util.ConfigUtils
+import java.io.File
+
+abstract class AbstractConfig<T>(
+        name: String,
+        protected val filePath: String
+) : SettingMultiGroup(name), IConfig<T> {
+
+    override val file get() = File("$filePath$name.json")
+    override val backup get() =  File("$filePath$name.bak")
+
+    override fun save() {
+        File(filePath).run {
+            if (!exists()) mkdirs()
+        }
+
+        saveToFile(this, file, backup)
+    }
+
+    override fun load() {
+        try {
+            loadFromFile(this, file)
+        } catch (e: Exception) {
+            KamiMod.LOG.warn("Failed to load latest, loading backup.")
+            loadFromFile(this, backup)
+        }
+    }
+
+    /**
+     * Save a group to a file
+     *
+     * @param group Group to save
+     * @param file Main file of [group]'s json
+     * @param backup Backup file of [group]'s json
+     */
+    protected fun saveToFile(group: SettingGroup, file: File, backup: File) {
+        ConfigUtils.fixEmptyJson(file)
+        ConfigUtils.fixEmptyJson(backup)
+        if (file.exists()) file.copyTo(backup, true)
+
+        file.bufferedWriter().use {
+            gson.toJson(group.write(), it)
+        }
+    }
+
+    /**
+     * Load settings values of a group
+     *
+     * @param group Group to load
+     * @param file file of [group]'s json
+     */
+    protected fun loadFromFile(group: SettingGroup, file: File) {
+        ConfigUtils.fixEmptyJson(file)
+
+        val jsonObject = file.bufferedReader().use {
+            parser.parse(it).asJsonObject
+        }
+        group.read(jsonObject)
+    }
+
+    /**
+     * Contains a gson object and a parser object
+     */
+    protected companion object {
+        val gson: Gson = GsonBuilder().setPrettyPrinting().create()
+        val parser = JsonParser()
+    }
+
+}
