+++ src/main/java/me/zeroeightsix/kami/setting/configs/IConfig.kt
@@ -1,6 +1,6 @@
-package me.zeroeightsix.kami.setting
+package me.zeroeightsix.kami.setting.configs
 
-import me.zeroeightsix.kami.setting.settings.AbstractSetting
+import me.zeroeightsix.kami.setting.settings.SettingRegister
 import java.io.File
 
 /**
@@ -8,23 +8,13 @@ import java.io.File
  *
  * @param T Type to have extension function for registering setting
  */
-interface IFinalGroup<T> {
+interface IConfig<T> : SettingRegister<T> {
 
     /** Main file of the config */
     val file: File
 
     /** Backup file of the config */
     val backup: File
-
-    /**
-     * Register a setting to this group
-     *
-     * @param S Type of the setting
-     * @param setting Setting to register
-     *
-     * @return [setting]
-     */
-    fun <S : AbstractSetting<*>> T.setting(setting: S): S
 
     /**
      * Save this group to its .json file
